Refactor iceberg hadoop and hive catalog tests
