+++ presto-iceberg/src/test/java/com/facebook/presto/iceberg/hive/TestIcebergDistributedHive.java
@@ -11,8 +11,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.iceberg;
+package com.facebook.presto.iceberg.hive;
 
+import com.facebook.presto.iceberg.TestAbstractIcebergDistributed;
 import org.testng.annotations.Test;
 
 import static com.facebook.presto.iceberg.CatalogType.HIVE;
